requirements.txt
Flask==1.1.1
Flask-JSON==0.3.4
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
python-dotenv==0.10.3